# Conflicts:
#	Star/src/main/java/com/star/controller/BoardController.java
#	Star/src/main/resources/mappers/BoardMapper.xml
#	Star/src/main/resources/templates/star/userReport.html

+++ Star/src/main/java/com/star/domain/BoardDTO.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 
 @Getter
 @Setter
-public class BoardDTO {
+public class BoardDTO  {
 	
 	// 카테고리 테이블
 	private String category;
@@ -17,7 +17,7 @@ public class BoardDTO {
 	
 //	privat String category;		// 게시글 테이블의 카테고리도 카테고리 연결
 	
-	private int userNumber;		// 유저 테이블의 유저 넘버와 연결
+	private Long userNumber;		// 유저 테이블의 유저 넘버와 연결
 	
 	private String title;
